moved functions to global_path node_mock_global_path now depends on global_path.py
